Rename format_nodes to format
